Refactor EmailGenerator to consolidate email campaign creation into a guided 4-step process. Introduce new state management for campaign signal, persona, and pain points, while enhancing context analysis and email generation features. Remove EmailWizard component to streamline functionality and improve user experience without altering existing features.
